requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1

# Google Drive
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.128.0

# Document parsing
PyMuPDF==1.24.3
python-docx==1.1.0
striprtf==0.0.26

# Embeddings
pinecone==5.4.2

# OpenAI
openai==1.30.1

# Utilities
httpx==0.27.0
python-jose[cryptography]==3.3.0
tenacity==8.3.0
tiktoken==0.7.0
aiofiles==23.2.1
python-dotenv==1.0.1
